Add dockerfile for backend and docker-compose file
This change adds a Dockerfile for the backend which uses node as its
base image and installs all requirements and then runs the app forever.

It also adds a docker-compose file which runs the Dockerfile.
The .env-example includes the BACKEND_PORT variable which can be set but
defaults to 4200 if nothing is set